Add pyyaml to pytoml
I was using a pip version of pyyaml which I had fetched with pip earlier
in the day and not with poetry. Oops. Why? Pypi was returning SSL
certificate errors and poetry couldn't handle this. Pip how must talk to
a dependency cache as I fetched black, mypy, ipython, pytest and pyyaml
with no issues. Though I did forget to add them all to the pyproject
toml!